Added cookie-based checkout for users without accounts

+++ pages/cart.php
@@ -1,5 +1,3 @@
-<DOCTYPE html>
-<html>
   <?php
 	include "../php/functions.php";
 	// Set a few variables to empty in the event they don't get set later
@@ -21,7 +19,7 @@
 	
     session_start();
 
-  // From the tutorial
+  // From the tutorial but modified
   // If the user clicked the add to cart button on the product page we can check for the form data
   if (isset($_POST['product_name'], $_POST['quantity']) && is_string($_POST['product_name']) && is_numeric($_POST['quantity'])) {
     // Set the post variables so we easily identify them, also make sure they are integer
@@ -34,36 +32,70 @@
     $product_entry = mysqli_fetch_assoc($product);
     // Check if the product exists (array is not empty)
     if (isset($product_entry)) {
-		// Check for an existing quantity we need to increase
-		$check_cart = "SELECT * FROM shopping_cart WHERE username = '" . $_SESSION["currentUser"] . "'";
-		$query = $link->query($check_cart);
-		$updated_flag = false;
-		while ($item = mysqli_fetch_assoc($query)) {
-			if($item["product_name"] == $product_name) {
-				$updated_flag = true;
-				if ($_POST["updating"] == true) {
-					if ($quantity == 0) {
-						$remove_item = "DELETE FROM shopping_cart WHERE username = '" . $_SESSION["currentUser"] . "' AND product_name = '" . $product_name . "'";
-						// echo "<p>DEBUG: running query: " . $remove_item . "</p>";
-						$link->query($remove_item);
+		// If the user is logged in, use SQL so their cart persists indefinitely
+		if (isset($_SESSION['currentUser'])) {
+			// Check for an existing quantity we need to increase
+			$check_cart = "SELECT * FROM shopping_cart WHERE username = '" . $_SESSION["currentUser"] . "'";
+			$query = $link->query($check_cart);
+			$updated_flag = false;
+			while ($item = mysqli_fetch_assoc($query)) {
+				if($item["product_name"] == $product_name) {
+					$updated_flag = true;
+					if ($_POST["updating"] == true) {
+						if ($quantity == 0) {
+							$remove_item = "DELETE FROM shopping_cart WHERE username = '" . $_SESSION["currentUser"] . "' AND product_name = '" . $product_name . "'";
+							// echo "<p>DEBUG: running query: " . $remove_item . "</p>";
+							$link->query($remove_item);
+						}
+						else if ($quantity != $item["quantity"]) {
+							$change_quantity = "UPDATE shopping_cart SET quantity='" . $quantity . "' WHERE product_name = '" . $product_name . "'";
+							$link->query($change_quantity);
+						}
 					}
-					else if ($quantity != $item["quantity"]) {
+					else {
+						$quantity += $item["quantity"];
 						$change_quantity = "UPDATE shopping_cart SET quantity='" . $quantity . "' WHERE product_name = '" . $product_name . "'";
 						$link->query($change_quantity);
 					}
 				}
-				else {
-					$quantity += $item["quantity"];
-					$change_quantity = "UPDATE shopping_cart SET quantity='" . $quantity . "' WHERE product_name = '" . $product_name . "'";
-					$link->query($change_quantity);
-				}
+			}
+			// If it wasn't in the cart already, add it
+			if ($updated_flag == false) {
+				$add_product = "INSERT INTO shopping_cart (product_name, quantity, username) VALUES ('" . $product_name . "', '" . $quantity . "', '" . $_SESSION["currentUser"] . "')";
+				// Execute query
+				$link->query($add_product);
 			}
 		}
-		// If it wasn't in the cart already, add it
-		if ($updated_flag == false) {
-			$add_product = "INSERT INTO shopping_cart (product_name, quantity, username) VALUES ('" . $product_name . "', '" . $quantity . "', '" . $_SESSION["currentUser"] . "')";
-			// Execute query
-			$link->query($add_product);
+		// If the user isn't logged in, use cookies instead
+		else {
+			$cookie_name = cookieizeString($product_name);
+			// if the product is already in the cart
+			if (isset($_COOKIE[$cookie_name])) {
+				if ($_POST["updating"] == true) {
+					if ($quantity == 0) {
+						// Delete the cookie
+						setcookie($cookie_name, "", time() - 3600, "/");
+						unset($_COOKIE[$cookie_name]);
+					}
+					else if ($quantity != $_COOKIE[$product_name]) {
+						// Update cookie with new quantity and make it last 30 days
+						$quantity += $_COOKIE[$product_name];
+						setcookie($cookie_name, strval($quantity), time() + (86400 * 30), "/");
+						$_COOKIE[$cookie_name] = $quantity;
+					}
+				}
+				else {
+					$quantity += $_COOKIE[$product_name];
+					setcookie($cookie_name, strval($quantity), time() + (86400 * 30), "/");
+					$_COOKIE[$cookie_name] = $quantity;
+				}
+			}
+			// if the product is not in the cart
+			else {
+				setcookie($cookie_name, strval($quantity), time() + (86400 * 30), "/");
+				$_COOKIE[$cookie_name] = $quantity;
+			}
+			// Cookies corresponding to a product are named after a product and contain the quantity in the cart
 		}
 		// Prevent form resubmission...
 		// header('location: cart.php');
@@ -71,7 +103,9 @@
 	}
   }
   ?>
-  
+
+<DOCTYPE html>
+<html>
   <head>
     <title>A Brief Introduction to Coffee</title>
 	<link rel="stylesheet" href="../styles/main.css" />
@@ -88,35 +122,74 @@
     <main>
       <h2>Items</h2>
 	  <?php 
-	  	$get_cart = "SELECT product_name, quantity FROM shopping_cart WHERE username = '" . $_SESSION["currentUser"] . "'";
-		//echo "<p>DEBUG: running query: " . $get_cart . "</p>";
-		$cart = $link->query($get_cart);
-		//echo "<p>DEBUG: mysqli error: " . strval(mysqli_error($link)) . "</p>";
-		// echo "<p>DEBUG: products found: " . strval($cart) . "</p>";
 		$empty_flag = true;
-		while(($row = mysqli_fetch_assoc($cart)) != NULL) {
-			$empty_flag = false;
-			$query = mysqli_fetch_assoc($link->query("SELECT product_name, price, image_url, description FROM products WHERE product_name = '" . $row["product_name"] . "'"));
-			$price_str = sprintf("%.2f", $query['price']);
-			$item_total = $query['price'] * $row['quantity'];
-			$item_total_str = sprintf("%.2f", $item_total);
-			# Some of this is from the tutorial linked on the slides
-			echo<<<EOT
-			<div>
-				<h4>{$row['product_name']}</h4>
-				<figure class="storeFigure">
-					<img src='/images/{$query['image_url']}'></img>
-					<figcaption>$ {$price_str} × {$row['quantity']} = <strong>$ {$item_total_str}</strong></figcaption>
-				</figure>
-				<form class="shopForm" action="cart.php" method="post">
-					<input type="number" name="quantity" value="{$row['quantity']}" min="0" placeholder="Quantity" required>
-					<input type="hidden" name="product_name" value="{$row['product_name']}">
-					<input type="hidden" name="updating" value="true">
-					<input type="submit" value="Update">
-				</form>
+		// If the user is logged in
+		if(isset($_SESSION["currentUser"])) {
+			$get_cart = "SELECT product_name, quantity FROM shopping_cart WHERE username = '" . $_SESSION["currentUser"] . "'";
+			//echo "<p>DEBUG: running query: " . $get_cart . "</p>";
+			$cart = $link->query($get_cart);
+			//echo "<p>DEBUG: mysqli error: " . strval(mysqli_error($link)) . "</p>";
+			// echo "<p>DEBUG: products found: " . strval($cart) . "</p>";
+			$empty_flag = true;
+				while(($row = mysqli_fetch_assoc($cart)) != NULL) {
+					$empty_flag = false;
+					$query = mysqli_fetch_assoc($link->query("SELECT product_name, price, image_url, description FROM products WHERE product_name = '" . $row["product_name"] . "'"));
+					$price_str = sprintf("%.2f", $query['price']);
+					$item_total = $query['price'] * $row['quantity'];
+					$item_total_str = sprintf("%.2f", $item_total);
+					# Some of this is from the tutorial linked on the slides
+					echo<<<EOT
+					<div>
+						<h4>{$row['product_name']}</h4>
+						<figure class="storeFigure">
+							<img src='/images/{$query['image_url']}'></img>
+							<figcaption>$ {$price_str} × {$row['quantity']} = <strong>$ {$item_total_str}</strong></figcaption>
+						</figure>
+						<form class="shopForm" action="cart.php" method="post">
+							<input type="number" name="quantity" value="{$row['quantity']}" min="0" placeholder="Quantity" required>
+							<input type="hidden" name="product_name" value="{$row['product_name']}">
+							<input type="hidden" name="updating" value="true">
+							<input type="submit" value="Update">
+						</form>
 
-			</div>
-			EOT;
+					</div>
+					EOT;
+				}
+		}
+		else {
+			// Get a list of products and check if any corresponding cookies are set.
+			$get_inventory = "SELECT product_name, price, image_url, description FROM products";
+			//echo "<p>DEBUG: running query: " . $get_cart . "</p>";
+			$inventory = $link->query($get_inventory);
+			while(($row = mysqli_fetch_assoc($inventory)) != NULL) {
+				$cookie_name = cookieizeString($row["product_name"]);
+				if (isset($_COOKIE[$cookie_name])) {
+					$quantity = $_COOKIE[$cookie_name];
+					$price_str = sprintf("%.2f", $row['price']);
+					$item_total = $row['price'] * $_COOKIE[$cookie_name];
+					$item_total_str = sprintf("%.2f", $item_total);
+					# Some of this is from the tutorial linked on the slides
+					if ($quantity > 0) {
+						$empty_flag = false;
+						echo<<<EOT
+						<div>
+							<h4>{$row['product_name']}</h4>
+							<figure class="storeFigure">
+								<img src='/images/{$row['image_url']}'></img>
+								<figcaption>$ {$price_str} × {$quantity} = <strong>$ {$item_total_str}</strong></figcaption>
+							</figure>
+							<form class="shopForm" action="cart.php" method="post">
+								<input type="number" name="quantity" value="{$quantity}" min="0" placeholder="Quantity" required>
+								<input type="hidden" name="product_name" value="{$row['product_name']}">
+								<input type="hidden" name="updating" value="true">
+								<input type="submit" value="Update">
+							</form>
+
+						</div>
+						EOT;
+					}
+				}
+			}
 		}
 		// Only show checkout button if cart isn't empty
 		if ($empty_flag == false) {
@@ -128,6 +201,9 @@
 
 			EOT;
 		}
+	  else {
+		echo "<p>Empty cart.</p>";
+	  }
 	  ?>
     </main>
